pyproject.toml
[tool.poetry]
name = "t9_back"
version = "1.0.1"
description = "Back for t9"
authors = ["imflop <imflop@gmail.com>"]
exclude = ["tests", "*.tests"]

[tool.poetry.scripts]
t9-back = "t9-back.__main__:main"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
typer = "^0.4.0"
sentry-sdk = "^1.5.8"
alembic = "^1.7.7"
sqlmodel = "^0.0.6"
loguru = "^0.6.0"
orjson = "^3.6.7"
asyncpg = "^0.25.0"
PyJWT = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.0"}
qrcode = "^7.3.1"

[tool.poetry.dev-dependencies]
pytest-clarity = "^1.0.1"
pytest-test-groups = "^1.0.3"
pytest-cov = "^3.0.0"
pytest = "^7.1.1"
mypy = "^0.941"
isort = "^5.10.1"
black = "^22.1.0"
pylint = "^2.12.2"
pytest-asyncio = "^0.18.2"
flake8 = "^4.0.1"
ssort = "^0.11.5"
ipython = "^8.2.0"
requests = "^2.27.1"
pytest-faker = "^2.0.0"
pytest-postgres = "^0.7.0"
docker = "^5.0.3"
pytest-xdist = "^2.5.0"
factory-boy = "^3.2.1"
perflint = "^0.7.1"

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
lines_after_imports = 2
force_grid_wrap = 3

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
